Renamed threedi_toolbox.py into threedi_plugin.py
I did it before, but ran into some qgis error. Trying it again :-)
