+++ tests/test_threedi_plugin.py
@@ -1,6 +1,6 @@
-from ThreeDiToolbox import threedi_toolbox
+from ThreeDiToolbox import threedi_plugin
 
 
 def test_smoke():
     # We just import it. There used to be some import errors
-    assert threedi_toolbox
+    assert threedi_plugin
